Move Extensions on  object, string and Array to helper methods instead of extensions
