+++ src/Common/ISymbolInfoExtension.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using PlcInterface.Extensions;
 
 namespace PlcInterface;
 
@@ -17,7 +16,7 @@ internal static class ISymbolInfoExtension
     /// <param name="symbolHandler">A <see cref="ISymbolHandler"/> implementation.</param>
     /// <returns>A <see cref="IEnumerable{T}"/> of all child symbols.</returns>
     public static IEnumerable<ISymbolInfo> Flatten(this ISymbolInfo symbolInfo, ISymbolHandler symbolHandler)
-        => symbolInfo.ChildSymbols.Count == 0 ? symbolInfo.Yield() : symbolInfo.ChildSymbols.SelectMany(x => symbolHandler.GetSymbolinfo(x).Flatten(symbolHandler));
+        => symbolInfo.ChildSymbols.Count == 0 ? new[] { symbolInfo } : symbolInfo.ChildSymbols.SelectMany(x => symbolHandler.GetSymbolinfo(x).Flatten(symbolHandler));
 
     /// <summary>
     /// Flatten the type hierarchy.
@@ -30,7 +29,7 @@ internal static class ISymbolInfoExtension
     {
         if (symbolInfo.ChildSymbols.Count == 0)
         {
-            return (symbolInfo, value).Yield();
+            return new[] { (symbolInfo, value) };
         }
 
         return symbolInfo.ChildSymbols
@@ -40,7 +39,7 @@ internal static class ISymbolInfoExtension
                 object? childValue;
                 if (value is Array array)
                 {
-                    var indices = x.Name.GetIndices();
+                    var indices = IndicesHelper.GetIndices(x.Name);
                     childValue = array.GetValue(indices);
                 }
                 else
